restructure the scripts
